pyproject.toml
[project]
name = "doorstep"
version = "0.1.0"
description = "Venue playlist builder for Spotify"
authors = [
    {name = "Hannah Lyons",email = "jlyons@lexile.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "spotipy (>=2.23.0)",
    "python-dotenv (>=1.0.0)",
    "anthropic (>=0.84.0,<0.85.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "google-api-python-client (>=2.191.0,<3.0.0)",
    "google-auth-oauthlib (>=1.3.0,<2.0.0)"
]


[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
